% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxSLMADS1.R
\name{coxSLMADS1}
\alias{coxSLMADS1}
\title{Fit a cox proportional hazard Model (coxph) with pooling via Study Level Meta-Analysis (SLMA)}
\usage{
coxSLMADS1(formula, weights, data)
}
\arguments{
\item{formula}{a coxph() formula consistent with R syntax eg U~x+y+Z to regress
variables U on x, y and Z, where U is the survival object calculated separately using
the assign function ds.Surv.}

\item{weights}{a character string specifying a variable to be used as regression weights.
Specified in call to ds.coxSLMA.}

\item{data}{a character string specifying the name of a data.frame
holding the data to be analysed under the specified model.}
}
\value{
assesses and returns information about failure to pass disclosure traps
such as test of model complexity (saturation).
For more detailed information see help for ds.coxSLMA.
}
\description{
This is the first serverside aggregate function called by ds.coxSLMA. Fits a coxph model on data from single or multiple sources
with pooled co-analysis across studies being based on SLMA (Study Level Meta Analysis).
}
\details{
coxSLMADS1 is an aggregate function called by clientside function ds.coxSLMA.
ds.coxSLMA also calls another aggregate function coxSLMADS2. For more detailed
information see help for ds.coxSLMA.
}
\author{
Sofack, Ghislain.(Based on glmSLMADS1 by Paul Burton for DataSHIELD Development Team)
}
