% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvDS.R
\name{SurvDS}
\alias{SurvDS}
\title{Server side function called by \code{ds.Surv} to create the survival object}
\usage{
SurvDS(time, time2, event, type)
}
\arguments{
\item{time}{a numeric vector indicating the start or follow up time.}

\item{time2}{a numeric vector indicating the ending time of the interval for interval censored or counting process data only.}

\item{event}{a numeric vector providing the name of the status indicator(event).Usually binary
e.g 0=alive, 1=dead.For multiple endpoint data the event variable will be a factor, whose
first level is treated as censoring. Although unusual, the event indicator can be omitted,
in which case all subjects are assumed to have an event.}

\item{type}{character string specifying the type of censoring. Possible
values are "right", "left", "counting", "interval", "interval2" or "mstate"..}
}
\value{
the results of the survival object stored on the server.
}
\description{
This function is similar to R function \code{Surv}.
}
\details{
This function calculates the survival object
usually used as a response variable in the cox model formula..
}
\author{
Sofack, Ghislain. (based on corTestDS by Demetris Avraam, for DataSHIELD Development Team)
}
